Overhauls how authentication and endpoint work
Instead of relying only on a credential object, we split the old
credential object into two objects: os-credential and os-auth-token.

Now os-credential objects only stores credentials that are essential
to authenticate a user, while os-auth-token stores the token obtained
from Keystone after a successful authentication.

Note that the ability for an object to automatically request a new token
once the old one expires has been moved into the new os-auth-token.

Also, the main authentication function now returns two values, a
os-auth-token object that contains valid token and an endpoints, which
is an alist map that stores all the services currently available and
their endpoints.

This also means that we no longer "hardcoded" the base endpoint URL of each
API binding, but instead retrieve them dynamically from the Keystone
server that we are currently connecting to and store them into the
endpoint map. This should make the code more robust.

We also split the code that deals with authentication into its own file.