Adds a new macro, def-openstack-api, and uses it to define all API bindings
Finally, after weeks and months, I've finally managed to produce an
abstraction that I'm satisfied with: the def-openstack-api macro!

This new macro does all the heavy lifting when defining a new API
bindings, and greatly simplified the code base.

The key insight that lead to the creation of this macro is the
realization that the endpoints should be separated from the token
object, and that there should be another global special variable that
holds the alist map of active endpoints (and an accessor function to go
along with it).