Refactors get-value so that the keys are stored as a list
The reason for this refactor is because we need to be able to supply
variables as a key, which is very painful in the old version where all
the keys are all in one giant string delimited by a dot.

That the code became a lot more cleaner is also a nice bonus. Also, our
function is now tail-called, baby!